+++ amy-message-hub-api/.settings/org.eclipse.m2e.core.prefs
