+++ amy-message-hub/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>amy-message-bus</name>
+	<name>amy-message-hub</name>
 	<comment></comment>
 	<projects>
 	</projects>
