+++ core/src/main/java/de/unistuttgart/iaas/amyassist/amy/core/speech/recognizer/manager/RemoteSpeechRecognizerManager.java
@@ -31,7 +31,7 @@ import de.unistuttgart.iaas.amyassist.amy.core.speech.grammar.GrammarObjectsCrea
 import de.unistuttgart.iaas.amyassist.amy.core.speech.recognizer.handler.RecognitionResultHandler;
 import de.unistuttgart.iaas.amyassist.amy.core.speech.recognizer.handler.RemoteResultHandler;
 import de.unistuttgart.iaas.amyassist.amy.core.speech.tts.Output;
-import de.unistuttgart.iaas.amyassist.amy.messagebus.Broker;
+import de.unistuttgart.iaas.amyassist.amy.messagehub.MessageHub;
 
 /**
  * Class that manages the Recognizers belonging to a given AudioInputStream
@@ -51,13 +51,13 @@ public class RemoteSpeechRecognizerManager extends AbstractSpeechRecognizerManag
 	 *            Output Object where to Output the result of the Recognizer
 	 * @param grammarData
 	 *            DataSet of all GrammarObjects
-	 * @param broker
-	 *            Message Broker
+	 * @param messageHub
+	 *            Message Hub
 	 * 
 	 */
 	public RemoteSpeechRecognizerManager(AudioInputStream ais, SpeechInputHandler inputHandler, Output output,
-			GrammarObjectsCreator grammarData, Broker broker) {
-		super(ais, inputHandler, output, grammarData, broker);
+			GrammarObjectsCreator grammarData, MessageHub messageHub) {
+		super(ais, inputHandler, output, grammarData, messageHub);
 	}
 
 	/**
