+++ plugins/spotify/src/test/java/de/unistuttgart/iaas/amyassist/amy/plugin/spotify/PlayerlogicTest.java
@@ -37,8 +37,6 @@ import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.wrapper.spotify.model_objects.miscellaneous.CurrentlyPlayingContext;
 import com.wrapper.spotify.model_objects.miscellaneous.Device;
@@ -50,14 +48,14 @@ import com.wrapper.spotify.model_objects.specification.Track;
 import com.wrapper.spotify.model_objects.specification.User;
 
 import de.unistuttgart.iaas.amyassist.amy.core.di.annotation.Reference;
-import de.unistuttgart.iaas.amyassist.amy.messagebus.Broker;
+import de.unistuttgart.iaas.amyassist.amy.messagehub.MessageHub;
 import de.unistuttgart.iaas.amyassist.amy.plugin.spotify.entities.PlaylistEntity;
 import de.unistuttgart.iaas.amyassist.amy.plugin.spotify.logic.PlayerLogic;
 import de.unistuttgart.iaas.amyassist.amy.plugin.spotify.logic.Search;
 import de.unistuttgart.iaas.amyassist.amy.test.FrameworkExtension;
 import de.unistuttgart.iaas.amyassist.amy.test.TestFramework;
 
-@ExtendWith({ MockitoExtension.class, FrameworkExtension.class })
+@ExtendWith(FrameworkExtension.class)
 class PlayerLogicTest {
 
 	private static final String ID1 = "abc123";
@@ -79,20 +77,17 @@ class PlayerLogicTest {
 	private Paging<PlaylistSimplified> playlistsSpotifyFormat;
 	private List<PlaylistEntity> playlistsOwnFormat;
 
-	@Mock
 	private SpotifyAPICalls spotifyAPICalls;
 
-	@Mock
 	private Search search;
 
-	@Mock
-	private Broker broker;
+	private MessageHub broker;
 
 	@BeforeEach
 	public void init() {
 		this.spotifyAPICalls = this.testFramework.mockService(SpotifyAPICalls.class);
 		this.search = this.testFramework.mockService(Search.class);
-		this.broker = this.testFramework.mockService(Broker.class);
+		this.broker = this.testFramework.mockService(MessageHub.class);
 		this.playerLogic = this.testFramework.setServiceUnderTest(PlayerLogic.class);
 		initDevices();
 		initCurrentTrack();
