+++ pom.xml
@@ -18,6 +18,8 @@
 		<!-- Amy versions to depend on -->
 		<amy.di.api.version>0.5.0-Snapshot</amy.di.api.version>
 		<amy.di.version>0.5.0-Snapshot</amy.di.version>
+		<amy.message-hub.api.version>0.5.0-Snapshot</amy.message-hub.api.version>
+		<amy.message-hub.version>0.5.0-Snapshot</amy.message-hub.version>
 		<amy.api.version>0.5.0-Snapshot</amy.api.version>
 		<amy.plugin-manager-api.version>0.5.0-Snapshot</amy.plugin-manager-api.version>
 		<amy.http-server.version>0.5.0-Snapshot</amy.http-server.version>
@@ -59,7 +61,8 @@
 		<module>plugins/systemtime</module>
 		<module>plugins/weather</module>
 		<module>plugins/navigation</module>
-		<module>amy-message-bus</module>
+		<module>amy-message-hub-api</module>
+		<module>amy-message-hub</module>
 	</modules>
 	<licenses>
 		<license>
@@ -141,7 +144,12 @@
 			<dependency>
 				<groupId>de.unistuttgart.iaas.amyassist</groupId>
 				<artifactId>amy-di-api</artifactId>
-				<version>${amy.di.version}</version>
+				<version>${amy.di.api.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>de.unistuttgart.iaas.amyassist</groupId>
+				<artifactId>amy-message-hub-api</artifactId>
+				<version>${amy.message-hub.api.version}</version>
 			</dependency>
 			<dependency>
 				<groupId>de.unistuttgart.iaas.amyassist</groupId>
